Change table types length and precision

+++ Shiptech.Domain/Exceptions/InvalidAdditionValueException.cs
@@ -4,7 +4,7 @@ namespace Shiptech.Domain.Exceptions
 {
     public class InvalidAdditionValueException : ShiptechException
     {
-        public InvalidAdditionValueException(int? value) : base($"Invalid addition: given {value}, required X > 0")
+        public InvalidAdditionValueException(short? value) : base($"Invalid addition: given {value}, required X > 0")
         {
         }
     }
