+++ Shiptech.Domain/Exceptions/InvalidAssemblyLengthValueException.cs
@@ -4,7 +4,7 @@ namespace Shiptech.Domain.Exceptions
 {
     public class InvalidAssemblyLengthValueException : ShiptechException
     {
-        public InvalidAssemblyLengthValueException(int value) : base($"Invalid assembly length value: given {value}, required X > 0")
+        public InvalidAssemblyLengthValueException(short value) : base($"Invalid assembly length value: given {value}, required X > 0")
         {
         }
     }
