+++ Shiptech.Domain/Exceptions/InvalidD15IIValueException.cs
@@ -4,7 +4,7 @@ namespace Shiptech.Domain.Exceptions
 {
     public class InvalidD15IIValueException : ShiptechException
     {
-        public InvalidD15IIValueException(int? value) : base(
+        public InvalidD15IIValueException(short? value) : base(
             $"Invalid D15II value: given {value}, required X > 0 and X < 90")
         {
         }
