+++ Shiptech.Domain/Exceptions/InvalidD15IValueException.cs
@@ -4,7 +4,7 @@ namespace Shiptech.Domain.Exceptions
 {
     public class InvalidD15IValueException : ShiptechException
     {
-        public InvalidD15IValueException(int? value) : base(
+        public InvalidD15IValueException(short? value) : base(
             $"Invalid D15I value: given {value}, required X > 0 and X < 90")
         {
         }
