+++ Shiptech.Domain/Exceptions/InvalidDrawingLengthException.cs
@@ -4,7 +4,7 @@ namespace Shiptech.Domain.Exceptions
 {
     public class InvalidDrawingLengthException : ShiptechException
     {
-        public InvalidDrawingLengthException(int? value) : base($"Invalid drawing length: given {value}, required X > 0")
+        public InvalidDrawingLengthException(short? value) : base($"Invalid drawing length: given {value}, required X > 0")
         {
         }
     }
