+++ Shiptech.Domain/Exceptions/InvalidPrefabricationQuantityValueException.cs
@@ -4,7 +4,7 @@ namespace Shiptech.Domain.Exceptions
 {
     public class InvalidPrefabricationQuantityValueException : ShiptechException
     {
-        public InvalidPrefabricationQuantityValueException(int value) : base($"Invalid prefabrication quantity value: given {value}, required X > 0")
+        public InvalidPrefabricationQuantityValueException(short value) : base($"Invalid prefabrication quantity value: given {value}, required X > 0")
         {
         }
     }
