+++ Shiptech.Domain/ValueObjects/D15II.cs
@@ -4,9 +4,9 @@ namespace Shiptech.Domain.ValueObjects
 {
     public record D15II
     {
-        public int? Value { get; }
+        public short? Value { get; }
 
-        public D15II(int? value)
+        public D15II(short? value)
         {
             if (value is < 0 or > 90)
             {
@@ -16,7 +16,7 @@ namespace Shiptech.Domain.ValueObjects
             Value = value;
         }
 
-        public static implicit operator int?(D15II d15II) => d15II.Value;
-        public static implicit operator D15II(int? d15II) => new(d15II);
+        public static implicit operator short?(D15II d15II) => d15II.Value;
+        public static implicit operator D15II(short? d15II) => new(d15II);
     }
 }
